Add DistanceManager class
